requirements.txt
flask
requests
azure-identity
azure-storage-blob